TagReuirement modified refactor
